ELS 3.0.0 release
