+++ .gitignore
@@ -3,7 +3,6 @@
 # Compiled class file
 *.class
 
-# Log file
 *.log
 
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
