+++ .idea/runConfigurations/Hints_Remote_Publisher_backup.xml
@@ -5,7 +5,7 @@
     </envs>
     <option name="MAIN_CLASS_NAME" value="com.groksoft.els.Main" />
     <module name="ELS" />
-    <option name="PROGRAM_PARAMETERS" value="-h -k els-hints.keys -c debug -d debug --remote P -p meta/publisher.json -s meta/subscriber.json -T -i output/publisher-export.json -m output/publisher-backup-Mismatches.txt -W output/publisher-backup-WhatsNew.txt -f output/publisher-backup.log" />
+    <option name="PROGRAM_PARAMETERS" value="-k els-hints.keys -c debug -d debug --remote P -p meta/publisher.json -s meta/subscriber.json -T -i output/publisher-export.json -m output/publisher-backup-Mismatches.txt -W output/publisher-backup-WhatsNew.txt -f output/publisher-backup.log" />
     <option name="VM_PARAMETERS" value="-Dlog4j.configurationFile=file:$PROJECT_DIR$/lib/log4j2.xml" />
     <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$/../../../../plex/Plex/els" />
     <extension name="coverage" enabled="false" merge="false" sample_coverage="true" runner="idea" />
