+++ ELS.iml
@@ -4,8 +4,8 @@
     <exclude-output />
     <content url="file://$MODULE_DIR$">
       <sourceFolder url="file://$MODULE_DIR$/src" isTestSource="false" />
-      <excludeFolder url="file://$MODULE_DIR$/artifacts/vio" />
       <excludeFolder url="file://$MODULE_DIR$/mock/TestRun" />
+      <excludeFolder url="file://$MODULE_DIR$/src/Status-Server-META-INF" />
     </content>
     <orderEntry type="inheritedJdk" />
     <orderEntry type="sourceFolder" forTests="false" />
