+++ artifacts/complete/README.txt
@@ -12,6 +12,7 @@ users get started with ELS.
  * linux subdirectory : Example Linux bash scripts.
  * meta subdirectory : Example publisher and subscriber library files and targets file.
  * windows subdirectory : Example Windows batch script.
+ * els-hints.keys : Example name/key file for ELS hints.
 
 ELS has many options that may be used in different combinations. Only a
 few are shown in these example files.
