+++ src/META-INF/MANIFEST.MF
@@ -1,3 +1,2 @@
 Manifest-Version: 1.0
 Main-Class: com.groksoft.els.Main
-
